+++ packages/pigeon/tool/shared/generation.dart
@@ -98,6 +98,9 @@ Future<int> generateTestPigeons({required String baseDir}) async {
     int generateCode = await runPigeon(
       input: './pigeons/$input.dart',
       dartOut: '$sharedDartOutputBase/lib/src/generated/$input.gen.dart',
+      dartTestOut: input == 'message'
+          ? '$sharedDartOutputBase/test/test_message.gen.dart'
+          : null,
       // Android
       kotlinOut: skipLanguages.contains(GeneratorLanguage.kotlin)
           ? null