+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.2+1
+
+* Updates `README` with a usage section and fix app-facing interface documentation.
+
+## 0.0.2
+
+* Adds Android implementation.
+
+## 0.0.1+3
+
+* Fixes the pub badge source.
+
 ## 0.0.1+2
 
 * Bumps Android's androidx.annotation:annotation dependency from 1.5.0 to 1.8.0.